requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-core
langchain-groq
llmsherpa
reportlab
markdown
weasyprint
